pyproject.toml
[project]
name = "adaptive_sdk"
version = "0.1.0b1"
description = "Python SDK for Adaptive Engine"
authors = [
  { name = "Vincent Debergue", email = "vincent@adaptive-ml.com" },
  { name = "Joao Moura", email = "joao@adaptive-ml.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["httpx >= 0.27.2", "gql >= 3.5.0", "pydantic[email] >= 2.9.0", "pyhumps >= 3.8.0"]

[project.optional-dependencies]
dev = ["ariadne-codegen >= 0.14.0"]

[tool.ariadne-codegen]
schema_path = "src/adaptive_sdk/schema.graphql"
queries_path = "src/adaptive_sdk/queries.graphql"
target_package_name = "graphql_client"
target_package_path = "src/adaptive_sdk"
client_name = "GQLClient"
async_client = false
opentelemetry_client = true
enable_custom_operations = true

[tool.ariadne-codegen.scalars.IdOrKey]
type = "str"
[tool.ariadne-codegen.scalars.Timestamp]
type = "int"
[tool.ariadne-codegen.scalars.InputDatetime]
type = "int | str"
[tool.ariadne-codegen.scalars.TimestampSec]
type = "int"
